[skip ci] Merge "Handle oddly sized disk images" am: 852cf66d0d am: 71f990724a am: bf37f1ca14 am: 6e20f3008a
Original change: https://android-review.googlesource.com/c/device/google/cuttlefish/+/1486303

GitOrigin-RevId: e3104681a2cd627352fcff47074d96d27317dea3

+++ staging/common/libs/utils/size_utils.cpp
@@ -20,8 +20,8 @@
 
 namespace cuttlefish {
 
-uint32_t AlignToPowerOf2(uint32_t val, uint8_t align_log) {
-  uint32_t align = 1 << align_log;
+uint64_t AlignToPowerOf2(uint64_t val, uint8_t align_log) {
+  uint64_t align = 1ULL << align_log;
   return ((val + (align - 1)) / align) * align;
 }
